+++ obj/Debug/netstandard2.1/MyPluginInfo.cs
@@ -3,7 +3,7 @@ namespace BepInEx5.PluginTemplate
     public static class MyPluginInfo
     {
         public const string PLUGIN_GUID = "AlwaysHearWalkie";
-        public const string PLUGIN_NAME = "My first plugin";
-        public const string PLUGIN_VERSION = "1.3.0";
+        public const string PLUGIN_NAME = "Always Hear Active Walkies";
+        public const string PLUGIN_VERSION = "1.4.0";
     }
 }